identify/requirements.txt
oauth2>=1.9.0,<2.0.0
absl-py>=0.4.1,<0.5.0
tweepy>=3.6.0,<3.7.0
PyYAML>=3.13.0,<3.14.0
google-cloud-datastore>=1.7.0,<1.8.0
google-cloud-logging>=1.6.0,<1.7.0
google-cloud-error-reporting>=0.30.0,<0.31.0
google-cloud-pubsub>=0.37.2,<0.38.0
google-cloud-storage>=1.12.0,<1.13.0
psq>=0.7.0,<0.8.0
Flask>=1.0.2,<1.1.0
more-itertools>=4.3.0,<4.4.0
gunicorn>=19.9.0,<20.0.0
honcho>=1.0.1,<1.1.0
dataclasses>=0.6.0,<0.7.0
mecab-python3>=0.7.0,<0.8.0
nltk>=3.3.0,<3.4.0
gensim>=3.5.0,<3.6.0
tqdm>=4.26.0,<4.27.0
pandas>=0.23.4,<0.24.0
scikit-learn>=0.19.2,<0.20.0
joblib>=0.12.4,<0.13.0
wordcloud>=1.5.0,<1.6.0
matplotlib>=2.2.3,<2.2.4
